change float character stats to int

+++ Assets/Scripts/Characters/CaveP.cs
@@ -7,14 +7,14 @@ namespace Characters
 
     public class CaveP : MonoBehaviour
     {
-        private const float HP = 10;
-        private const float MP = 5;
-        private const float SPE = 8;
-        private const float RAN = 6;
-        private const float DMG = 10;
-        private const float MDMG = 1;
-        private const float DEF = 6;
-        private const float MDEF = 2;
+        private const int HP = 10;
+        private const int MP = 5;
+        private const int SPE = 8;
+        private const int RAN = 6;
+        private const int DMG = 10;
+        private const int MDMG = 1;
+        private const int DEF = 6;
+        private const int MDEF = 2;
         private const int LVL = 1;
         
         private void Awake()
