+++ Assets/Scripts/Characters/Knight.cs
@@ -6,14 +6,14 @@ namespace Characters
 {
     public class Knight : MonoBehaviour
     {
-        private const float HP = 10;
-        private const float MP = 6;
-        private const float SPE = 1;
-        private const float RAN = 5;
-        private const float DMG = 6;
-        private const float MDMG = 2;
-        private const float DEF = 10;
-        private const float MDEF = 8;
+        private const int HP = 10;
+        private const int MP = 6;
+        private const int SPE = 1;
+        private const int RAN = 5;
+        private const int DMG = 6;
+        private const int MDMG = 2;
+        private const int DEF = 10;
+        private const int MDEF = 8;
         private const int LVL = 1;
         
         private void Awake()
